autoplot.basis {ggfortify}	R Documentation

Autoplot spline basis instances

Description

Autoplot spline basis instances

Usage

## S3 method for class 'basis'
autoplot(object, data, n = 256, ...)

Arguments

object
spline basis object

data
x-values at which to evaluate the splines. Optional. By default, an evenly spaced sequence of 256 values covering the range of the splines will be used.

n
If data is not provided, instead use an evenly-spaced sequence of x-values of this length (plus one, since both endpoints are included). If data is provided, this argument is ignored.

...
Ignored.

Value

ggplot

Examples

library(splines)
x <- seq(0, 1, by=0.001)
spl <- bs(x, df=6)
autoplot(spl)
autoplot(spl, n=5)

[Package ggfortify version 0.4.4 Index]