fortify.aareg {ggfortify}	R Documentation

Convert survival::aareg to data.frame

Description

Convert survival::aareg to data.frame

Usage

## S3 method for class 'aareg'
fortify(model, data = NULL, maxtime = NULL,
  surv.connect = TRUE, melt = FALSE, ...)

Arguments

model
survival::aareg instance

data
original dataset, if needed

maxtime
truncate the input to the model at time "maxtime"

surv.connect
logical frag indicates whether connects survival curve to the origin

melt
Logical flag indicating whether to melt each timeseries as variable

...
other arguments passed to methods

Value

data.frame

Examples

library(survival)
fortify(aareg(Surv(time, status) ~ age + sex + ph.ecog, data = lung, nmin = 1))
fortify(aareg(Surv(time, status) ~ age + sex + ph.ecog, data = lung, nmin = 1), melt = TRUE)

[Package ggfortify version 0.4.3 Index]