geom_confint {ggfortify}	R Documentation

Connect observations by stairs.

Description

Connect observations by stairs.

Usage

geom_confint(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, ...)

Arguments

mapping
the aesthetic mapping

data
a layer specific dataset

stat
the statistical transformation to use on the data for this layer

position
the position adjustment to use for overlapping points on this layer

na.rm
logical frag whether silently remove missing values

...
other arguments passed to methods

[Package ggfortify version 0.4.4 Index]