get.dtindex.continuous {ggfortify}	R Documentation

Get Date vector continue to ts index

Description

Get Date vector continue to ts index

Usage

get.dtindex.continuous(data, length, is.tsp = FALSE, is.date = NULL)

Arguments

data
ts instance

length
A number to continue

is.tsp
Logical frag whether data is tsp itself or not

is.date
Logical frag indicates whether the stats::ts is date or not. If not provided, regard the input as date when the frequency is 4 or 12.

Value

vector

Examples

ggfortify:::get.dtindex.continuous(AirPassengers, length = 10)
ggfortify:::get.dtindex.continuous(UKgas, length = 10)

[Package ggfortify version 0.4.2 Index]