ggfreqplot {ggfortify}	R Documentation

Plot seasonal subseries of time series, generalization of stats::monthplot

Description

Plot seasonal subseries of time series, generalization of stats::monthplot

Usage

ggfreqplot(data, freq = NULL, nrow = NULL, ncol = NULL,
  conf.int = FALSE, conf.int.colour = "#0000FF",
  conf.int.linetype = "dashed", conf.int.fill = NULL,
  conf.int.alpha = 0.3, conf.int.value = 0.95, facet.labeller = NULL, ...)

Arguments

data
stats::ts instance

freq
Length of frequency. If not provided, use time-series frequency

nrow
Number of plot rows

ncol
Number of plot columns

conf.int
Logical flag indicating whether to plot confidence intervals

conf.int.colour
line colour for confidence intervals

conf.int.linetype
line type for confidence intervals

conf.int.fill
fill colour for confidence intervals

conf.int.alpha
alpha for confidence intervals

conf.int.value
Coverage probability for confidence interval

facet.labeller
A vector used as facet labels

...
Keywords passed to autoplot.ts

Value

ggplot

Examples

ggfreqplot(AirPassengers)
ggfreqplot(AirPassengers, freq = 4)
ggfreqplot(AirPassengers, conf.int = TRUE)

[Package ggfortify version 0.4.4 Index]