is_derived_from {ggfortify}	R Documentation

Check object is target class, or object is data.frame fortified from target.

Description

Check object is target class, or object is data.frame fortified from target.

Usage

is_derived_from(object, target)

Arguments

object
instance to be checked. For data.frame, check whether it is fortified from target class

target
class name

Value

logical

Examples

ggfortify:::is_derived_from(prcomp(iris[-5]), 'prcomp')

[Package ggfortify version 0.4.2 Index]